Fixed several compiler warnings.
Changed the generation number in a PdfReference to pdf_uint16 from unsigned long. This might save a few bytes for every object 
depending on the padding done by the compiler. The range of a generation number is defined to be 0-65535 so a 16bit unsigned 
integer is enough to store a generation number.