Enable read-only access to streams.
It appears a bug may be exposed in PoDoFo's delayed loading of streams where the object's contents are re-read even if they've already been delay-loaded. This causes stored pointers to the old objects to be invalid (violating the "document tree is only modified through the model" assumption we make here). A crash can be triggered by clicking on an object with a stream, then accessing a grandchild node of it.

Don't be surprised if podofobrowser crashes sometimes until this bug in podofo is fixed.