ADDED: If PODOFO_USE_UNORDERED_MAP is defined during compilation std::tr1::unordered_map will be used in PdfDictionary.
I was curious after Craigs mail today, wether using an unordered_map would make any difference in PoDoFo. 

Some benchmarking tests using callgrind have shown
that unordered_map is a little faster for writing and AddKey
but of course slower for GetKey and HasKey. As PdfDictionaries
are usually very small the difference for GetKey and HasKey is
not very large and should therefore be rarely noticeable.

By default this define is not set and std::map will be used.