Added an exception for linking OpenSSL as suggested by Dan Horak.
This was found by a review for inclusion into Fedora.

the reviewer found a legal issue that I missed and it is the incompatibility of
LGPLv2 (its clause #10) and GPLv2 (clause #6) used for podofo
library/tools and OpenSSL that is linked with podofo. It can be solved
by adding an exception clause to the GPL/LGPL used for podofo, see
http://www.gnome.org/~markmc/openssl-and-the-gpl.html for an example.