Lame workarounds for compilation under bcc32
For some reason, C++Builder XE2, when compiling PoDoFo as DLL,
doesn't set PdfDifferenceEncoding virtual functions in PdfEncoding
parent, thus for example the call to PdfEncoding::IsAutoDelete()
on a PdfDifferenceEncoding object can fail in runtime with
"Pure virtual function called" message and an abort of the whole
application. As I've no idea what causes this (yet), I use this
lame workarounds for now.

The PdfInfo change is funny, XE2 thought the '//' is a comment
begin, regardless it's inside a string constant.