Introduce PdfMemDocument::IsLoaded()
When an encrypted file fails to be loaded due to incorrect
or none password provided, then the PdfMemDocument is left
in a state which can lead to a crash (when for example
trying to access AcroForm). The added IsLoaded() is to check
whether the load was properly completed.

As a side-effect, a bug with not freed (but not leaked) parser
at PdfMemDocument::SetPassword() had been fixed.