ADDED: Patch by Matthew Brincke: Documentation cleanup for "ErrOk on success" return value where none is given
Clean up documentation for methods with return type "void" and a constructor
where a return of "ErrOk" is still mentioned: remove such mentions, also
clarify some parameter and method documentation and correct the constructor's
documentation. Also remove of @returns in favor of \returns and correction of
return-value documentation (both in PdfVariant.h).