Patch by Jan-Erik S
The following diff contains a change needed to make podofo 0.9.5
able to parse an old PDF-1.3 file without crashing in font creation.

The fix is small and contained, but a bit crude, so you will want to review it.

The file contains the following sequence of object (font is referenced
at 37 0 and the crash is triggered by the line "/W [ 0 35 0 R ]"):

35 0 obj
[ 250 276 247 298 168 776 671 671 549 713
604 609 725 660 665 653 666 693 845 900
759 832 988 812 782 842 892 417 328 877
394 587 667 250 ]
endobj
36 0 obj
<<
/Type /Font
/BaseFont /Fd5656
/CIDSystemInfo <<
/Ordering (Identity)
/Registry (Adobe)
/Supplement 0
>>
/FontDescriptor 34 0 R
/FontInfo <<
/FSType 4
>>
/Subtype /CIDFontType0
/W [ 0 35 0 R ]
>>
endobj
37 0 obj
<<
/Type /Font
/BaseFont /Fd5656-Identity-H
/Bezier 1
/DescendantFonts [ 36 0 R ]
/Encoding /Identity-H
/Subtype /Type0
/ToUnicode 20 0 R
>>