Fixed issue #56: SEGV in podofoimpose caused by PdfTranslator::migrateResource(PdfObject*) NULL dereference
This adds NULL checks in the appropriate places and should have been part of
the fix for issue #45 in the previous commit. NULL being returned outside of
a loop means that nothing can be done in that call of migrateResource anymore,
then NULL is returned from that call also, immediately. That NULL check has
been added by me after the publishing of the provisional patch yesterday (UTC)
and hasn't been tested with a fresh test case (other than the PoC, I mean) and
only tested with the older compiler versions.